Refactor nvhttp display scale handlers
