+++ cmake/compile_definitions/common.cmake
@@ -98,6 +98,10 @@ set(SUNSHINE_TARGET_FILES
         "${CMAKE_SOURCE_DIR}/src/webhook.h"
         "${CMAKE_SOURCE_DIR}/src/nvhttp.cpp"
         "${CMAKE_SOURCE_DIR}/src/nvhttp.h"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/display_scale.cpp"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/display_scale.h"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/url_utils.cpp"
+        "${CMAKE_SOURCE_DIR}/src/nvhttp/url_utils.h"
         "${CMAKE_SOURCE_DIR}/src/nvhttp_stream_start.cpp"
         "${CMAKE_SOURCE_DIR}/src/nvhttp_stream_start.h"
         "${CMAKE_SOURCE_DIR}/src/abr.cpp"
