+++ src/nvhttp/url_utils.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <string>
+
+namespace nvhttp::url_utils {
+
+  std::string
+  decode(std::string value);
+
+}  // namespace nvhttp::url_utils
